refactored the Barnes0Hut algortithms to be subclasses. Moved the non-performant versions to 'tests' so they can still be used to compare results. Removed the userData arg from the ForceObject
 Changes to be committed:
	deleted:    jung-algorithms/src/main/java/edu/uci/ics/jung/layout/algorithms/FRBHLayoutAlgorithm.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/algorithms/FRBHVisitorLayoutAlgorithm.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/algorithms/FRLayoutAlgorithm.java
	deleted:    jung-algorithms/src/main/java/edu/uci/ics/jung/layout/algorithms/SpringBHLayoutAlgorithm.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/algorithms/SpringBHVisitorLayoutAlgorithm.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/algorithms/SpringLayoutAlgorithm.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/spatial/BarnesHutQuadTree.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/spatial/ForceObject.java
	modified:   jung-algorithms/src/main/java/edu/uci/ics/jung/layout/spatial/Node.java
	new file:   jung-algorithms/src/test/java/edu/uci/ics/jung/layout/algorithms/FRBHIteratorLayoutAlgorithm.java
	new file:   jung-algorithms/src/test/java/edu/uci/ics/jung/layout/algorithms/SpringBHIteratorLayoutAlgorithm.java
	modified:   jung-algorithms/src/test/java/edu/uci/ics/jung/layout/spatial/FRLayoutsTest.java
	modified:   jung-algorithms/src/test/java/edu/uci/ics/jung/layout/spatial/FRLayoutsTimingTest.java
	modified:   jung-algorithms/src/test/java/edu/uci/ics/jung/layout/spatial/SpringLayoutsTest.java
	modified:   jung-samples/src/main/java/edu/uci/ics/jung/samples/ShowLayouts.java
