+++ jung-algorithms/src/main/java/edu/uci/ics/jung/layout/spatial/Node.java
@@ -1,6 +1,5 @@
 package edu.uci.ics.jung.layout.spatial;
 
-import java.util.Optional;
 import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -129,7 +128,7 @@ public class Node<T> {
     SE = new Node(x + width, y + height, width, height);
   }
 
-  public void visit(ForceObject<T> target, Optional userData) {
+  public void visit(ForceObject<T> target) {
     if (this.forceObject == null || target.element.equals(this.forceObject.element)) {
       return;
     }
@@ -143,7 +142,7 @@ public class Node<T> {
             target.getElement(),
             target.p);
       }
-      target.addForceFrom(this.forceObject, userData);
+      target.addForceFrom(this.forceObject);
       log.trace("added force from {} so its now {}", this.forceObject, target);
     } else {
       // not a leaf
@@ -164,15 +163,15 @@ public class Node<T> {
               target.getElement(),
               target.p);
         }
-        target.addForceFrom(this.forceObject, userData);
+        target.addForceFrom(this.forceObject);
         log.trace("added force from {} so its now {}", this.forceObject, target);
 
       } else {
         // down the tree we go
-        NW.visit(target, userData);
-        NE.visit(target, userData);
-        SW.visit(target, userData);
-        SE.visit(target, userData);
+        NW.visit(target);
+        NE.visit(target);
+        SW.visit(target);
+        SE.visit(target);
       }
     }
   }
