+++ jung-algorithms/src/test/java/edu/uci/ics/jung/layout/algorithms/SpringBHIteratorLayoutAlgorithm.java
@@ -15,26 +15,29 @@ import edu.uci.ics.jung.layout.model.Point;
 import edu.uci.ics.jung.layout.spatial.BarnesHutQuadTree;
 import edu.uci.ics.jung.layout.spatial.ForceObject;
 import java.util.ConcurrentModificationException;
+import java.util.Iterator;
 import java.util.function.Function;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * This subclass of SpringLayoutAlgorithm applies a Barnes-Hut QuadTree optimization during the
- * calculation of node repulsion
+ * calculation of node repulsion. It uses an Iterator over the reduced set of nodes for comparison.
+ * As it is not as performant as the Visitor model, it is in the test module for comparison of
+ * output between the implementations.
  *
  * @author Tom Nelson
  */
-public class SpringBHVisitorLayoutAlgorithm<N> extends SpringLayoutAlgorithm<N>
+public class SpringBHIteratorLayoutAlgorithm<N> extends SpringLayoutAlgorithm<N>
     implements IterativeContext {
 
-  private static final Logger log = LoggerFactory.getLogger(SpringBHVisitorLayoutAlgorithm.class);
+  private static final Logger log = LoggerFactory.getLogger(SpringBHIteratorLayoutAlgorithm.class);
 
   private BarnesHutQuadTree<N> tree;
 
-  public SpringBHVisitorLayoutAlgorithm() {}
+  public SpringBHIteratorLayoutAlgorithm() {}
 
-  public SpringBHVisitorLayoutAlgorithm(
+  public SpringBHIteratorLayoutAlgorithm(
       Function<? super EndpointPair<N>, Integer> length_function) {
     super(length_function);
   }
@@ -56,7 +59,6 @@ public class SpringBHVisitorLayoutAlgorithm<N> extends SpringLayoutAlgorithm<N>
 
     try {
       for (N node : graph.nodes()) {
-
         if (layoutModel.isLocked(node)) {
           continue;
         }
@@ -65,37 +67,39 @@ public class SpringBHVisitorLayoutAlgorithm<N> extends SpringLayoutAlgorithm<N>
         if (svd == null) {
           continue;
         }
-        ForceObject<N> nodeForceObject =
-            new ForceObject(node, layoutModel.apply(node)) {
-              @Override
-              protected void addForceFrom(ForceObject other) {
+        double dx = 0, dy = 0;
 
-                if (other == null || node == other.getElement()) {
-                  return;
-                }
-                Point p = this.p;
-                Point p2 = other.p;
-                if (p == null || p2 == null) {
-                  return;
-                }
-                double vx = p.x - p2.x;
-                double vy = p.y - p2.y;
-                double distanceSq = p.distanceSquared(p2);
-                if (distanceSq == 0) {
-                  f = f.add(random.nextDouble(), random.nextDouble());
-                } else if (distanceSq < repulsion_range_sq) {
-                  double factor = 1;
-                  f = f.add(factor * vx / distanceSq, factor * vy / distanceSq);
-                }
-              }
-            };
-        tree.visit(nodeForceObject);
-        Point f = nodeForceObject.f;
-        double dlen = f.x * f.x + f.y * f.y;
+        ForceObject<N> nodeForceObject = new ForceObject<>(node, layoutModel.apply(node));
+        Iterator<ForceObject<N>> forceObjectIterator =
+            new BarnesHutQuadTree.ForceObjectIterator<>(tree, nodeForceObject);
+        while (forceObjectIterator.hasNext()) {
+          ForceObject<N> nextForceObject = forceObjectIterator.next();
+          if (nextForceObject == null || node == nextForceObject.getElement()) {
+            continue;
+          }
+          Point p = nodeForceObject.p;
+          Point p2 = nextForceObject.p;
+          if (p == null || p2 == null) {
+            continue;
+          }
+          double vx = p.x - p2.x;
+          double vy = p.y - p2.y;
+          double distanceSq = p.distanceSquared(p2);
+          if (distanceSq == 0) {
+            dx += random.nextDouble();
+            dy += random.nextDouble();
+          } else if (distanceSq < repulsion_range_sq) {
+            double factor = 1;
+            dx += factor * vx / distanceSq;
+            dy += factor * vy / distanceSq;
+          }
+        }
+
+        double dlen = dx * dx + dy * dy;
         if (dlen > 0) {
           dlen = Math.sqrt(dlen) / 2;
-          svd.repulsiondx += f.x / dlen;
-          svd.repulsiondy += f.y / dlen;
+          svd.repulsiondx += dx / dlen;
+          svd.repulsiondy += dy / dlen;
         }
       }
     } catch (ConcurrentModificationException cme) {
